simple_robot/requirements.txt
websocket-client==1.8.0
sproto==0.1.3
locust==2.32.1
websockets==14.1